Fix issue with tiles
The problem was that sometimes it was not detecting the colliders, because it was starting exactly on the surface.  Not sure why sometimes it worked or not, maybe numerical issues.  Just made it start a bit higher.
Also using a layer for the domino surface, so it doesn't try to place one piece on top of another